# services/contracts/market.openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "KebuMarket + Store Builder — retro 8.4b", "version": "0.1.0", "description": "Canonical Zod → OpenAPI. Java truth catalog/inventory/store + idempotent checkout (Idempotency-Key). FE mfe-kebumarket + mfe-store-builder via api-client." },
  "servers": [{ "url": "http://localhost:4000" }],
  "paths": {
    "/market/catalog": { "get": { "summary": "Catalog paginated", "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string", "nullable": true } }], "responses": { "200": { "description": "Catalog", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogResponse" } } } } } } },
    "/market/checkout": { "post": { "summary": "Idempotent checkout", "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutRequest" } } } }, "responses": { "200": { "description": "Order", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutResponse" } } } }, "409": { "description": "Duplicate Idempotency-Key different payload" } } } },
    "/store/{id}": { "get": { "summary": "Store by id/slug", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Store", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Store" } } } } } } },
    "/store/{id}/inventory": { "get": { "summary": "Inventory", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Inventory", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InventoryResponse" } } } } } } }
  },
  "components": { "schemas": {
    "Product": { "type": "object", "required": ["id", "title", "priceMinor", "currency", "stock"], "properties": { "id": { "type": "string" }, "title": { "type": "string" }, "description": { "type": "string" }, "priceMinor": { "type": "integer" }, "currency": { "type": "string", "minLength": 3 }, "imageUrl": { "type": "string", "format": "uri", "nullable": true }, "stock": { "type": "integer" }, "isShowcase": { "type": "boolean" } } },
    "CatalogResponse": { "type": "object", "required": ["data", "nextCursor"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }, "nextCursor": { "type": "string", "nullable": true } } },
    "CartItem": { "type": "object", "required": ["productId", "qty", "priceMinor", "title"], "properties": { "productId": { "type": "string" }, "qty": { "type": "integer", "minimum": 1, "maximum": 99 }, "priceMinor": { "type": "integer" }, "title": { "type": "string" } } },
    "CheckoutRequest": { "type": "object", "required": ["items", "idempotencyKey"], "properties": { "items": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/CartItem" } }, "idempotencyKey": { "type": "string", "format": "uuid" } } },
    "CheckoutResponse": { "type": "object", "required": ["orderId", "status", "totalMinor"], "properties": { "orderId": { "type": "string" }, "status": { "type": "string", "enum": ["created", "paid", "failed"] }, "totalMinor": { "type": "integer" } } },
    "Store": { "type": "object", "required": ["id", "name", "slug"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "slug": { "type": "string", "pattern": "^[a-z0-9-]+$" }, "description": { "type": "string" }, "status": { "type": "string", "enum": ["draft", "published", "archived"] }, "isShowcase": { "type": "boolean" } } },
    "InventoryItem": { "type": "object", "required": ["id", "storeId", "productId", "title", "priceMinor", "currency", "stock"], "properties": { "id": { "type": "string" }, "storeId": { "type": "string" }, "productId": { "type": "string" }, "title": { "type": "string" }, "priceMinor": { "type": "integer" }, "currency": { "type": "string" }, "stock": { "type": "integer" } } },
    "InventoryResponse": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/InventoryItem" } } } }
  } }
}